CatchRecord method change long to Double for dashboard and report which can still open if the catch record don't have data

+++ src/main/java/baitmate/ImplementationMethod/CatchRecordServiceImpl.java
@@ -49,7 +49,7 @@ public class CatchRecordServiceImpl implements CatchRecordService {
     }
 
     @Override
-    public long calculateAverageCatchesPerUser() {
+    public Double calculateAverageCatchesPerUser() {
         return catchRecordRepository.calculateAverageCatchesPerUser();
     }
